+++ src/main/java/Controller/ProfileImage/DeleteProfileImage.java
@@ -1,7 +1,11 @@
 package Controller.ProfileImage;
 
+import Annotation.RequiresPermission;
 import Controller.Auth.BaseServlet;
+import Enums.Permissions;
+import Repository.AuthRepository;
 import Service.ProfileImageService;
+import Utill.GetUserIdOfCurrentLogin;
 import Utill.JsonResponse;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -13,8 +17,10 @@ import java.util.Map;
 public class DeleteProfileImage extends BaseServlet {
 
     private final ProfileImageService userImageService = new ProfileImageService();
+    private final AuthRepository authRepository = new AuthRepository();
 
     @Override
+    @RequiresPermission(Permissions.DELETE_PROFILE_IMAGE)
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
 
@@ -26,6 +32,14 @@ public class DeleteProfileImage extends BaseServlet {
         }
 
         int userId = Integer.parseInt(pathInfo.substring(1)); // remove "/"
+        int loginUserId= GetUserIdOfCurrentLogin.getUserIdFromRequest(req);
+
+        String role=authRepository.getRoleByUserId(loginUserId);
+
+        if (role!="ADMIN" && loginUserId!=userId){
+            JsonResponse.forbidden(resp, "You don't have permission to do this task");
+            return;
+        }
 
         boolean deleted = userImageService.deleteProfileImage(userId);
